requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dnspython==2.2.1
Flask==2.0.3
Flask-PyMongo==2.3.0
fonttools==4.31.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
kiwisolver==1.4.1
MarkupSafe==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
pymongo==4.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-string-utils==1.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tenacity==8.0.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.0.3
